+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed download speed after the download is completed.
 - Changed the database downloads table columns to include save and temp paths of each download.
 - Separated the downloader connections wiring logic into a different function.
+- Changed database file path to appdata directory.
 
 ## [v1.4.2]  29-08-2026